Fix viewport value list
From the Chrome error log:
Viewport argument value "device-width;" for key "width" is invalid, and has been ignored. Note that ';' is not a separator in viewport values. The list should be comma-separated.
Viewport argument value "device-height;" for key "height" is invalid, and has been ignored. Note that ';' is not a separator in viewport values. The list should be comma-separated.
Viewport argument value "no;" for key "user-scalable" is invalid, and has been ignored. Note that ';' is not a separator in viewport values. The list should be comma-separated